Theme C: collapse three CLAP embedders onto a shared _ClapBase
The clap, clap_general, and clap_music audio embedders were ~90%
identical, differing only by checkpoint id, display name, and progress
labels. Extract a Flask-free _ClapBase in _clap_shared.py (matching the
existing _Dinov2Base pattern); the three variant modules become thin
subclasses overriding name/display_name/label/model_id (and is_default
for the baseline). All three now also share the baseline's resampling-JIT
warmup, which general/music previously lacked.
